Add retired_at timestamp for Rubygems
Some gems despite retaining non-yanked versions need to be marked
as retired or deprecated so that people understand the dependencies
they are using are no longer maintained.

This could be used by the RubyGem client or Bundler when people use a
specific warning level (not by default, of course) to let them know that
they are relying on unmaintened dependencies and that it may be time to
update to another gem or vendor the gem to apply patches.

I don't believe this something we read from gemspecs since it would be
quite odd to ask a maintainer for a gem that is no longer maintained to
release a final version. That would also be confusing for users.

Instead this should be a feature accessible via the API and the web
interface and I plan to implement those two as the next step for this feature.

A further idea worth considering is allow gem maintainers to direct users
to a third-party gem that solves the same problem but is actively maintained.